+++ packages/xrpl/src/client/connection.ts
@@ -421,6 +421,7 @@ export class Connection extends EventEmitter {
    * @returns A promise that resolves to void when the connection is fully established.
    * @throws Error if the websocket initialized is somehow null.
    */
+  // eslint-disable-next-line max-lines-per-function -- Many error code conditionals to check.
   private async onceOpen(connectionTimeoutID: NodeJS.Timeout): Promise<void> {
     if (this.ws == null) {
       throw new XrplError('onceOpen: ws is null')
@@ -449,20 +450,30 @@ export class Connection extends EventEmitter {
       this.ws.removeAllListeners()
       this.ws = null
 
-      if(code === undefined) {
-        console.error(`Disconnected but the disconnect code was undefined (The given reason was ${reason}).` +
-        `This could be caused by an exception being thrown during a 'connect' callback. ` +
-        `Disconnecting with code 1011 to indicate an internal error has occurred.`)
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- undefined can be passed in.
+      if (code === undefined) {
+        // eslint-disable-next-line no-console -- The error is helpful for debugging.
+        console.error(
+          `Disconnected but the disconnect code was undefined (The given reason was ${reason.toString()}).` +
+            `This could be caused by an exception being thrown during a 'connect' callback. ` +
+            `Disconnecting with code 1011 to indicate an internal error has occurred.`,
+        )
 
-        // Error code 1011 represents an Internal Error according to
-        // https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent/code
-        this.emit('disconnected', 1011)
+        /*
+         * Error code 1011 represents an Internal Error according to
+         * https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent/code
+         */
+        const internalErrorCode = 1011
+        this.emit('disconnected', internalErrorCode)
       } else {
         this.emit('disconnected', code)
       }
 
-      // If this wasn't a manual disconnect, then lets reconnect ASAP.
-      // Code can be undefined if there's an exception while connecting.
+      /*
+       * If this wasn't a manual disconnect, then lets reconnect ASAP.
+       * Code can be undefined if there's an exception while connecting.
+       */
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- undefined can be passed in.
       if (code !== INTENTIONAL_DISCONNECT_CODE && code !== undefined) {
         this.intentionalDisconnect()
       }